client/requirements.txt
numpy==1.18.5
Pillow==7.2.0
requests==2.24.0